JVM_IR: move SuspendLambdaLowering to a separate file
